+++ pyudev/device.py
@@ -366,6 +366,32 @@ class Device(Mapping):
         return ensure_unicode_string(libudev.udev_device_get_sysname(self))
 
     @property
+    def sys_number(self):
+        """
+        The trailing number of the :attr:`sys_name` as unicode string, or
+        ``None``, if the device has no trailing number in its name.
+
+        .. note::
+
+           The number is returned as unicode string to preserve the exact
+           format of the number, especially any leading zeros:
+
+           >>> device = pyudev.Device.from_path(ctx, '/sys/devices/LNXSYSTM:00')
+           >>> device.sys_number
+           u'00'
+
+           To work with numbers, explicitly convert them to ints:
+
+           >>> int(device.sys_number)
+           0
+
+        .. versionadded:: 0.11
+        """
+        number = libudev.udev_device_get_sysnum(self)
+        if number is not None:
+            return ensure_unicode_string(number)
+
+    @property
     def device_type(self):
         """
         Device type as unicode string, or ``None``, if the device type is